Modificacion de documento & mostrarDatoEnID(DONE)
